Better positioning
Positioning should be done only when requested.  This allows the caller
to leave the options.position blank and style via CSS without being
overriden at the tag level e.g. .sigma-tooltips {position: absolute:
left/top: … }, which is useful when tooltips are large it’s best to pin
them to fixed location on the viewport