Edge curvature
This patch enhances the library’s edge curvature capabilities, which
are at present hard-coded.

Of particular interest are cases when multiple edges exist between two
nodes.  the current framework displays edges atop one another so what
is needed is a way to specify curvature for individual edges.

the patch to sigma.utils.getQuadraticControlPoint() adds a fifth
parameter which contains a collection of coefficients dictating the
control point’s position.

patches to all calling contexts have been made to provide the
coefficients as found on the edge, or to pick up the global settings
value ’curvatureCoefficients’.  if neither exists the current default
is used.

additionally, a method sigma.canvas.edges.autoCurve() has been provided
to automatically curve edges for nodes with multiple connections.  this
method is accessible as an overload to the sigma.canvas.edges.curve()
method.

finally, the parallelEdges module overrides the functionality of the
getQuadraticControlPoint() function which is a pernicious approach to
solve the same problem and has been re-implemented as
getQuadraticControlPoint2 by calling the original method, with all
calls to it from the same module patched