Insert glyphs canvas before mouse canvas
The last canvas in the HTML renderer may not always be the mouse canvas, (for example if the legends canvas is added before the glyphs render function is called).  If the glyphs canvas is placed after the mouse canvas, all mouse interactions break.  This places the glyph canvas before the mouse canvas regardless of where it was initially inserted in the stack.